ci: add nightly post-quantum benchmarks

+++ crypto-ffi/bindings/jvm/src/jmh/kotlin/com/wire/benchmark/AddUser.kt
@@ -23,7 +23,7 @@ open class AddUser {
         "5",
         "7"
     )
-    var cipherSuite: UShort = 1u
+    var cipherSuite: Int = 1
 
     @Param("1", "10", "100")
     var userCount: Int = 0
@@ -35,7 +35,7 @@ open class AddUser {
     @Setup(Level.Invocation)
     fun setup() {
         runBlocking {
-            val cipherSuite = CipherSuite.entries.first { it.value == cipherSuite }
+            val cipherSuite = CipherSuite.entries.first { it.value.toInt() == cipherSuite }
             val options = CcInitOptions(CcInitOptions.Mode.WithBasicCredential(cipherSuite))
             aliceCc = ccInit(options)
             conversationId = createConversation(aliceCc)
